# Dual-row Optimizations on top of KeccakRound gate