--- a/kimchi/src/circuits/polynomials/keccak/mod.rs
+++ b/kimchi/src/circuits/polynomials/keccak/mod.rs
@@ -8,7 +8,7 @@ pub const QUARTERS: usize = 4;
 pub const ROUNDS: usize = 24;
 pub const RATE: usize = 1088 / 8;
 pub const CAPACITY: usize = 512 / 8;
-pub const KECCAK_COLS: usize = 1965;
+pub const KECCAK_COLS: usize = 1100;
 
 use crate::circuits::expr::constraints::ExprOps;
 use ark_ff::PrimeField;