… book/src/specs/kimchi.md
… kimchi/src/circuits/polynomials/keccak/gadget.rs
… kimchi/src/circuits/polynomials/keccak/witness.rs
… kimchi/src/linearization.rs
… kimchi/src/plonk_sponge.rs
… kimchi/src/proof.rs
… kimchi/src/prover.rs
… kimchi/src/verifier.rs
… kimchi/src/verifier_index.rs